+++ src/Presentation/SmartStore.Web/Scripts/public.product-detail.js
@@ -160,8 +160,6 @@
 			var self  = this;
 			var opts = this.options;
 
-			console.log(opts.galleryHeight);
-
 			gallery = $('#pd-gallery').smartGallery({
 			    height: opts.galleryHeight,
 				enableDescription: opts.showImageDescription,