+++ changelog.md
@@ -35,7 +35,7 @@
 * GMC feed: Supporting fields multipack, bundle, adult, energy efficiency class and custom label (0 to 4)
 * #760 Setting to set a customer role for new registered users
 * #800 Multi-store: Option to display all orders of all stores for customer in frontend
-* #457 Added option to hide the default image for categories
+* #457 Added option to hide the default image for categories and products
 
 ### Improvements
 * (Perf) Implemented static caches for URL aliases and localized properties. Increases app startup and request speed by up to 30%.